Implemented Fast Maximum Filter and fixed Gaussian blur.

+++ src/logic/data_structures/kernel.rs
@@ -1,13 +1,13 @@
 pub struct Kernel1D {
-    pub weights: Vec<i32>,
+    pub weights: Vec<f64>,
     pub divisor_factor: f64,
 }
 
 impl Kernel1D {
-    pub fn new(weights: Vec<i32>, divisor: i32) -> Kernel1D {
+    pub fn new(weights: Vec<f64>, divisor: f64) -> Kernel1D {
         return Kernel1D {
             weights,
-            divisor_factor: 1.0 / divisor as f64
+            divisor_factor: 1.0 / divisor
         };
     }
